Fix tailwind-variants with Vue
Followed cva Vue example for typing VariantProps. It is more involved as
the SFC compiler does not like complex types, not conditionals, nor
Pick, or key in, etc. I tried it all.

Opted for tailwind-variants as it allows getting `defaultVariants` from
the object, cva did not.

I did not resolve the fact that a manual reload is needed when default
props are changed in variants.